+++ bin/ncp/CONFIG/nc-nextcloud.sh
@@ -83,9 +83,6 @@ EOF
 
 configure()
 {
-  ## IF BETA SELECTED ADD "pre" to DOWNLOAD PATH
-  [[ "$BETA" == yes ]] && local PREFIX="pre"
-
   ## DOWNLOAD AND (OVER)WRITE NEXTCLOUD
   cd /var/www/
 
@@ -213,8 +210,7 @@ EOF
 Description = Push daemon for Nextcloud clients
 
 [Service]
-Environment=PORT=7867
-Environment=NEXTCLOUD_URL=https://localhost
+Environment = PORT=7867 # Change if you already have something running on this port
 ExecStart = /path/to/push/binary/notify_push /path/to/nextcloud/config/config.php
 User=www-data