+++ bin/ncp/UPDATES/unattended-upgrades.sh
@@ -34,6 +34,7 @@ Unattended-Upgrade::Origins-Pattern {
 o=Debian,n=$RELEASE;
 o=deb.sury.org,n=$RELEASE;
 o="Raspberry Pi Foundation",n=$RELEASE;
+o="Raspbian",n=$RELEASE;
 }
 Dpkg::Options {
    "--force-confdef";